// src/app/components/admin/role/action-enum.enum.ts
export enum ActionEnum{
        INDEX=100,
        POST=101,
        PUT=102,
        DELETE=103,
        
        CreateUser = 1,

        ViewAllRequestDetails = 2,

        AssigningRequestToStaffMembers = 3,

        ReviewRequestAndCommentsAndRequestingAnupdate = 4,

        ReviewUploadingFilesAndRequestingAnUpdate = 5,

        ReviewAllPayments = 6,

        FillDownloadAndSendServiceOfAgreement = 7,

        ReviewServiceOfAgreementSubmittedByConsumer = 8,

        DownloadCertificate = 9,

        SendAuditPlan = 10,

        ViewConsumerNoteAboutAuditPlan = 11,

        EditAuditPlan = 12,

        TakeDecisionIfThereCasesOfNonConformityOrNo = 13,

        FillFormOfCasesOfNonConformityD5SFDAHA14 = 14,

        FillReportOfCasesOfNonConformityD5SFDAHA61 = 15,


        ReviewCorrectiveActionPlanSubmittedByTheConsumer = 16,

        ReturnOrAcceptCorrectiveActionPlanSubmittedByTheConsumer = 17,

        FillAuditReportAndSendToDecisionCommittee = 18,

        TakeDecisionWithWritingComment= 19,
}