// src/app/components/admin/role/page-enum.enum.ts
export enum PageEnum{
        Requests = 1,

        Pages = 2,

        Consumer = 3,

        User = 4,

        Payment = 5,

        Employee = 6,

        Country = 7,

        Agreement = 8,

        AuditPlan = 9,

        DecisionCommittee = 10,

        Certificate = 11,

        Configration = 12,

        Dashboard = 13,
        Role=14,
        Job=15,
        City=16
}